Add support for zz_api photo creation and upload.  Refactor before filters to be called inline

+++ app/controllers/albums_controller.rb
@@ -3,31 +3,38 @@
 #
 
 class AlbumsController < ApplicationController
+  # NOTE: this controller has been converted to use the new return unless filter style
+  # of calling explicitly in each controller method.
+  #
+  #
+  #
   # methods that should be in except lists since they are open or manged by an only
   # the list was getting out of hand having to include twice
   # this is the common set the is excepted from require_user and require_album
-  def self.except_methods
-    return [
-        :index, :invalidate_cache, :zz_api_albums,
-        :my_albums_json, :liked_albums_json, :my_albums_public_json, :liked_albums_public_json, :liked_users_public_albums_json,
-        :zz_api_my_albums_json, :zz_api_liked_albums_json, :zz_api_my_albums_public_json, :zz_api_liked_albums_public_json, :zz_api_liked_users_public_albums_json,
-        :invited_albums_json, :zz_api_invited_albums_json
-    ]
-  end
-
-  before_filter :require_user,              :except => except_methods + [ :download ]
-  before_filter :require_same_user_json,    :only =>   [ :my_albums_json, :liked_albums_json, :zz_api_my_albums_json, :zz_api_liked_albums_json,
-      :invited_albums_json, :zz_api_invited_albums_json, :invalidate_cache
-  ]
-  before_filter :require_album,             :except => except_methods + [ :create, :new ]
-  before_filter :require_album_admin_role,  :only =>   [ :destroy, :edit, :update, :add_group_members ]
+  #def self.except_methods
+  #  return [
+  #      :index, :invalidate_cache, :zz_api_albums,
+  #      :my_albums_json, :liked_albums_json, :my_albums_public_json, :liked_albums_public_json, :liked_users_public_albums_json,
+  #      :zz_api_my_albums_json, :zz_api_liked_albums_json, :zz_api_my_albums_public_json, :zz_api_liked_albums_public_json, :zz_api_liked_users_public_albums_json,
+  #      :invited_albums_json, :zz_api_invited_albums_json
+  #  ]
+  #end
+  #
+  #before_filter :require_user,              :except => except_methods + [ :download ]
+  #before_filter :require_same_user_json,    :only =>   [ :my_albums_json, :liked_albums_json, :zz_api_my_albums_json, :zz_api_liked_albums_json,
+  #    :invited_albums_json, :zz_api_invited_albums_json, :invalidate_cache
+  #]
+  #before_filter :require_album,             :except => except_methods + [ :create, :new ]
+  #before_filter :require_album_admin_role,  :only =>   [ :destroy, :edit, :update, :add_group_members ]
 
   # displays the "Select album type screen" used in the wizard
   def new
+    return unless require_user
     render :layout => false
   end
 
   def create
+    return unless require_user
     if params[:album_type].nil?
       render :text => "Error No Album Type Supplied. Please Choose Album Type.", :status=>500 and return
     end
@@ -44,6 +51,7 @@ class AlbumsController < ApplicationController
   # of what the album email will be as the album name changes
   # @album is set in the require_album before filter
   def preview_album_email
+    return unless require_user && require_album && require_album_admin_role
     if base = params[:album][:name]
       begin
         if @album.name_unique?( base )
@@ -74,18 +82,21 @@ class AlbumsController < ApplicationController
   # displays the name album page for the wizard
   # @album is set by the require album before filter
   def name_album
+    return unless require_user && require_album && require_album_admin_role
     render :layout => false
   end
 
   # displays the album privacy page for the wizard
   # @album is set by the require album before filter
   def privacy
+    return unless require_user && require_album && require_album_admin_role
     render :layout => false
   end
 
   # displays the edit page for the wizard
   # @album is set by the require album before filter
   def edit
+    return unless require_user && require_album && require_album_admin_role
     @photos = @album.photos
     render :layout => false
   end
@@ -93,6 +104,7 @@ class AlbumsController < ApplicationController
   # updates @album attributes
   # @album is set by the require_album before_filter
   def update
+    return unless require_user && require_album && require_album_admin_role
     # The AlbumCoverPicker sends an empty string when no cover has been selected. Delete param if that is the case
     # TODO: Fix AlbumCoverPicker to omit cover_photo_id parameter if not set by user
     params[:album].delete(:cover_photo_id) if params[:album][:cover_photo_id] && params[:album][:cover_photo_id].length <= 0
@@ -109,6 +121,7 @@ class AlbumsController < ApplicationController
   # returns JSON used to populate
   # the "Group" tab
   def edit_group
+    return unless require_user && require_album && require_album_admin_role
     hash = {
         :user => {
           :has_facebook_token => !current_user.identity_for_facebook.credentials_valid?.nil?,
@@ -137,6 +150,7 @@ class AlbumsController < ApplicationController
   # change a group member from contributor to viewer
   # or vice-versa
   def update_group_member
+    return unless require_user && require_album && require_album_admin_role
     user_id = params[:member][:id]
 
     if params[:member][:permission] == 'contributor'
@@ -151,6 +165,7 @@ class AlbumsController < ApplicationController
 
   # remove member from group
   def delete_group_member
+    return unless require_user && require_album && require_album_admin_role
     user_id = params[:member][:id]
     @album.acl.remove_user(user_id)
     render :json => ''
@@ -158,6 +173,7 @@ class AlbumsController < ApplicationController
 
 
   def add_group_members
+    return unless require_user && require_album && require_album_admin_role
 
     emails,errors = Share.validate_email_list(  params[:emails] )
     if errors.length > 0
@@ -190,50 +206,15 @@ class AlbumsController < ApplicationController
   end
 
   def group_members
+    return unless require_user && require_album && require_album_admin_role
     render :json => get_group_members
   end
 
 
-  # fetches the album paths, this is common code shared
-  # between zz_api and normal apis
-  # returns @loader, and @session_loader context which
-  # can be used to create the specific form of data that
-  # each type of call wants
-  def get_albums(user, zz_api)
-    # determine if we should be fetching the view based on public or private data
-    user_is_me = current_user?(user)
-    private_view = user_is_me || (!current_user.nil? && current_user.support_hero?)
-    public = !private_view
-
-    # preload the expected cache data
-    loader = Cache::Album::Manager.shared.make_loader(user, public)
-    loader.pre_fetch_albums
-
-    # call the following methods to get the json paths for my_albums, my_liked_albums, etc
-    # The url paths returned are based on whether we are viewing ourselves or somebody else (based on the public flag)
-    #
-    # The paths are relative to the host so start with /service/...
-    session_loader = nil
-
-    # When showing the view for a user who is not the current user we
-    # fetch the public information.  However, if we have a current user
-    # and that user likes and can see one or more of the other users
-    # albums we must merge the view on the client by checking to see
-    # if any of the albums the current user likes belong to the viewed user
-    # we do this by returning the url to fetch liked_albums for the session
-    # user
-    if public && !current_user.nil?
-      # ok, we have a valid session user and we are viewing somebody else so pull in our liked_albums
-      session_loader = Cache::Album::Manager.shared.make_loader(current_user, false)
-      session_loader.pre_fetch_albums
-    end
-
-    @loader = loader
-    @session_loader = session_loader
-  end
 
   # This is effectively the users homepage
   def index
+    return unless require_nothing
     begin
       @user = User.find(params[:user_id])
     rescue ActiveRecord::RecordNotFound => e
@@ -256,6 +237,7 @@ class AlbumsController < ApplicationController
 
 
   def zz_api_albums()
+    return unless require_nothing
     zz_api do
       user = User.find(params[:user_id])
 
@@ -264,14 +246,14 @@ class AlbumsController < ApplicationController
       session_loader = @session_loader
       if !session_loader.nil?
         session_liked_users_albums = session_loader.liked_users_album_loader.current_version_key
-        session_liked_users_albums_path = zz_api_liked_users_albums_path(session_loader)
+        session_liked_users_albums_link = zz_api_liked_users_albums_link(session_loader)
         session_invited_albums = session_loader.invited_album_loader.current_version_key
-        session_invited_albums_path = zz_api_invited_albums_path(session_loader)
+        session_invited_albums_link = zz_api_invited_albums_link(session_loader)
       else
         session_liked_users_albums = nil
-        session_liked_users_albums_path = nil
+        session_liked_users_albums_link = nil
         session_invited_albums = nil
-        session_invited_albums_path = nil
+        session_invited_albums_link = nil
       end
 
       album_meta = {
@@ -279,23 +261,24 @@ class AlbumsController < ApplicationController
         :logged_in_user_id              => current_user.nil? ? nil : current_user.id,
         :public                         => loader.public,
         :my_albums                      => loader.my_album_loader.current_version_key,
-        :my_albums_path                 => zz_api_my_albums_path(loader),
+        :my_albums_path                 => zz_api_my_albums_link(loader),
         :liked_albums                   => loader.liked_album_loader.current_version_key,
-        :liked_albums_path              => zz_api_liked_albums_path(loader),
+        :liked_albums_path              => zz_api_liked_albums_link(loader),
         :liked_users_albums             => loader.liked_users_album_loader.current_version_key,
-        :liked_users_albums_path        => zz_api_liked_users_albums_path(loader),
+        :liked_users_albums_path        => zz_api_liked_users_albums_link(loader),
         :session_user_liked_albums      => session_liked_users_albums,
-        :session_user_liked_albums_path => session_liked_users_albums_path,
+        :session_user_liked_albums_path => session_liked_users_albums_link,
         :invited_albums                 => loader.invited_album_loader.current_version_key,
-        :invited_albums_path            => zz_api_invited_albums_path(loader),
+        :invited_albums_path            => zz_api_invited_albums_link(loader),
         :session_user_invited_albums      => session_invited_albums,
-        :session_user_invited_albums_path => session_invited_albums_path
+        :session_user_invited_albums_path => session_invited_albums_link
       }
     end
   end
 
   # invalidate the current cache for this user - essentially a forced cache flush version change
   def invalidate_cache
+    return unless require_same_user_json
     user_id = params[:user_id]
     Cache::Album::Manager.shared.user_invalidate_cache(user_id)
     render :json => ''
@@ -310,9 +293,9 @@ class AlbumsController < ApplicationController
     url << "?ver=#{ver}"
   end
 
-  def zz_api_my_albums_path(loader, force_zero_ver = false)
+  def zz_api_my_albums_link(loader, force_zero_ver = false)
     ver = force_zero_ver ? 0 : loader.my_album_loader.current_version_key
-    url = loader.public ? zz_api_my_albums_public_json_path(loader.user_id) : zz_api_my_albums_json_path(loader.user_id)
+    url = loader.public ? zz_api_my_albums_public_path(loader.user_id) : zz_api_my_albums_path(loader.user_id)
     url << "?ver=#{ver}"
   end
 
@@ -322,9 +305,9 @@ class AlbumsController < ApplicationController
     url << "?ver=#{ver}"
   end
 
-  def zz_api_liked_albums_path(loader, force_zero_ver = false)
+  def zz_api_liked_albums_link(loader, force_zero_ver = false)
     ver = force_zero_ver ? 0 : loader.liked_album_loader.current_version_key
-    url = loader.public ? zz_api_liked_albums_public_json_path(loader.user_id) : zz_api_liked_albums_json_path(loader.user_id)
+    url = loader.public ? zz_api_liked_albums_public_path(loader.user_id) : zz_api_liked_albums_path(loader.user_id)
     url << "?ver=#{ver}"
   end
 
@@ -333,9 +316,9 @@ class AlbumsController < ApplicationController
     liked_users_public_albums_json_path(loader.user_id) + "?ver=#{ver}"
   end
 
-  def zz_api_liked_users_albums_path(loader, force_zero_ver = false)
+  def zz_api_liked_users_albums_link(loader, force_zero_ver = false)
     ver = force_zero_ver ? 0 : loader.liked_users_album_loader.current_version_key
-    zz_api_liked_users_public_albums_json_path(loader.user_id) + "?ver=#{ver}"
+    zz_api_liked_users_public_albums_path(loader.user_id) + "?ver=#{ver}"
   end
 
   def invited_albums_path(loader, force_zero_ver = false)
@@ -343,9 +326,9 @@ class AlbumsController < ApplicationController
     invited_albums_json_path(loader.user_id) + "?ver=#{ver}"
   end
 
-  def zz_api_invited_albums_path(loader, force_zero_ver = false)
+  def zz_api_invited_albums_link(loader, force_zero_ver = false)
     ver = force_zero_ver ? 0 : loader.invited_album_loader.current_version_key
-    zz_api_invited_albums_json_path(loader.user_id) + "?ver=#{ver}"
+    zz_api_invited_albums_path(loader.user_id) + "?ver=#{ver}"
   end
 
   def albums_cache_setup(public)
@@ -373,10 +356,12 @@ class AlbumsController < ApplicationController
 
 # fetch my own albums
   def my_albums_json
+    return unless require_same_user_json
     my_albums_json_common(false)
   end
 
-  def zz_api_my_albums_json
+  def zz_api_my_albums
+    return unless require_same_user_json
     zz_api_self_render do
       my_albums_json_common(false)
     end
@@ -384,10 +369,12 @@ class AlbumsController < ApplicationController
 
 # fetch public albums for a given user
   def my_albums_public_json
+    return unless require_nothing
     my_albums_json_common(true)
   end
 
-  def zz_api_my_albums_public_json
+  def zz_api_my_albums_public
+    return unless require_nothing
     zz_api_self_render do
       my_albums_json_common(true)
     end
@@ -401,10 +388,12 @@ class AlbumsController < ApplicationController
 
 # fetch the albums I like
   def liked_albums_json
+    return unless require_same_user_json
     liked_albums_json_common(false)
   end
 
-  def zz_api_liked_albums_json
+  def zz_api_liked_albums
+    return unless require_same_user_json
     zz_api_self_render do
       liked_albums_json_common(false)
     end
@@ -412,10 +401,12 @@ class AlbumsController < ApplicationController
 
 # fetch the albums that a given user likes
   def liked_albums_public_json
+    return unless require_nothing
     liked_albums_json_common(true)
   end
 
-  def zz_api_liked_albums_public_json
+  def zz_api_liked_albums_public
+    return unless require_nothing
     zz_api_self_render do
       liked_albums_json_common(true)
     end
@@ -430,10 +421,12 @@ class AlbumsController < ApplicationController
   end
 
   def invited_albums_json
+    return unless require_same_user_json
     invited_albums_json_common
   end
 
-  def zz_api_invited_albums_json
+  def zz_api_invited_albums
+    return unless require_same_user_json
     zz_api_self_render do
       invited_albums_json_common
     end
@@ -449,10 +442,12 @@ class AlbumsController < ApplicationController
   end
 
   def liked_users_public_albums_json
+    return unless require_nothing
     liked_users_public_albums_json_common
   end
 
-  def zz_api_liked_users_public_albums_json
+  def zz_api_liked_users_public_albums
+    return unless require_nothing
     zz_api_self_render do
       liked_users_public_albums_json_common
     end
@@ -461,6 +456,7 @@ class AlbumsController < ApplicationController
 #deletes an album
 #@album is set by require_album before_filter
   def destroy
+    return unless require_user && require_album && require_album_admin_role
     # Album is found when the before filter calls authorized user
     if @album.destroy
       render :json => "Album deleted" and return
@@ -473,7 +469,8 @@ class AlbumsController < ApplicationController
 # we also have a watchdog sweeper that will
 # close batches with no new add activity after a 5 minute window
   def close_batch
-    album_id = params[:id]
+    return unless require_user && require_album && require_album_admin_role
+    album_id = @album.id
     if album_id
       UploadBatch.close_batch( current_user.id, album_id)
     end
@@ -482,11 +479,13 @@ class AlbumsController < ApplicationController
 
 #displays the "You have reached a password protected album, request access" dialog
   def pwd_dialog
+    return unless require_user && require_album
     render :layout => false
   end
 
 # Receives and processes a user's request for access into a password protected album
   def request_access
+    return unless require_user && require_album
     #TODO: Receive and process current_users request for access into the current album
   end
 
@@ -496,6 +495,8 @@ class AlbumsController < ApplicationController
   # plugin to nginx, so our job is to put together the list
   # of all photos that have been uploaded to amazon
   def download
+    return unless require_album
+
     unless  @album.can_user_download?( current_user )
       flash.now[:error] = "Only Authorized Album Group Members can download albums"
       if request.xhr?
@@ -545,28 +546,47 @@ class AlbumsController < ApplicationController
   end
 
   private
-#
-# To be run as a before_filter
-# Requires params[:id] to be present and be a valid album_id.
-# sets @album to be Album.find( params[:id ])
-# Throws ActiveRecord:RecordNotFound exception if params[:id] is not present or the album is not found
-  def require_album
-    begin
-      @album = Album.find( params[:id ])  #will trhow exception if params[:id] is not defined or album not found
-    rescue ActiveRecord::RecordNotFound => e
-      flash.now[:error] = "This operation requires an album, we could not find one because: "+e.message
-      if request.xhr?
-        head   :not_found
-      else
-        render :file => "#{Rails.root}/public/404.html", :status => 404
-      end
-      return false
+
+  # fetches the album paths, this is common code shared
+  # between zz_api and normal apis
+  # returns @loader, and @session_loader context which
+  # can be used to create the specific form of data that
+  # each type of call wants
+  def get_albums(user, zz_api)
+    # determine if we should be fetching the view based on public or private data
+    user_is_me = current_user?(user)
+    private_view = user_is_me || (!current_user.nil? && current_user.support_hero?)
+    public = !private_view
+
+    # preload the expected cache data
+    loader = Cache::Album::Manager.shared.make_loader(user, public)
+    loader.pre_fetch_albums
+
+    # call the following methods to get the json paths for my_albums, my_liked_albums, etc
+    # The url paths returned are based on whether we are viewing ourselves or somebody else (based on the public flag)
+    #
+    # The paths are relative to the host so start with /service/...
+    session_loader = nil
+
+    # When showing the view for a user who is not the current user we
+    # fetch the public information.  However, if we have a current user
+    # and that user likes and can see one or more of the other users
+    # albums we must merge the view on the client by checking to see
+    # if any of the albums the current user likes belong to the viewed user
+    # we do this by returning the url to fetch liked_albums for the session
+    # user
+    if public && !current_user.nil?
+      # ok, we have a valid session user and we are viewing somebody else so pull in our liked_albums
+      session_loader = Cache::Album::Manager.shared.make_loader(current_user, false)
+      session_loader.pre_fetch_albums
     end
+
+    @loader = loader
+    @session_loader = session_loader
   end
 
   def get_group_members
     group = []
-
 
     # collect contributors
     #
